refactor: divide pkg
